requirements.txt
PyPDF2
openai
python-dotenv
psycopg2-binary